Clean-up of the code
Changes to the code:
- Removed the Runge Kutta in energy direction. This did not work so
well in tests with a constant magnetic field and a flat injection
spectral index (gamma_inj=2.1).
- Combined the options in the parameter file into a more structured file
- Removed the old intensities and spectral indices without the
synchrotron spectrum integration
- Removed the

To be done:
- Include an option to read out the spectra at certain positions in the
parameter file
- Improve the accuracy of the grid points in frequency: nu1, nu2, nu3
and nu4
- Check that the adiabatic losses are working even if there is no model
for the radius
- Include a radius model for galaxies from the flux tube approximation
- Add Arpad’s graphical interface from python

+++ cosmic_rays.h
@@ -11,7 +11,7 @@ void adiabatic (void);
 
 struct grid_1d setup_initial_grid (void);
 
-void output_file (long i_max);
+void output_file (int i_max);
 
 double interpolate_frequency (double value, double value_low, double value_high, double nu, int jj);
